Split components into Layers and DrawTools
DrawMap no longer makes sense if it's not rendering the map. Also,
this folder was getting kind of big, so layers were split into a
different folder. MapFeatures could be moved into DrawTools because that
component can now reference a parent map in context.
