--- src/app/src/components/DrawMap/index.js
@@ -1,1 +1,0 @@
-export { default } from './DrawMap';
