+++ src/app/src/components/DrawTools/DeletePolygonConfirmModal.js
