+++ src/app/src/components/DrawTools/index.js
@@ -1,0 +1,1 @@
+export { default } from './DrawTools';
